+++ src/shared/state.ts
@@ -30,5 +30,6 @@ export function getDefaultState(): DesktopSettingsType {
     HTMLEmailAlwaysLoadRemoteContent: false,
     enableRelatedChats: false,
     galleryImageKeepAspectRatio: false,
+    useSystemUIFont: false,
   }
 }